+++ .gitignore
@@ -1,2 +1,3 @@
 Pipfile.lock
 Pipfile
+src/classes/oracle_data.json
